# expvar: optimize Map.String() and expvar HTTP handler