name         old time/op    new time/op    delta
MapString-4    2.69µs ± 2%    1.07µs ± 2%  -60.26%  (p=0.000 n=10+8)

name         old alloc/op   new alloc/op   delta
MapString-4      200B ± 0%      184B ± 0%   -8.00%  (p=0.000 n=10+10)

name         old allocs/op  new allocs/op  delta
MapString-4      8.00 ± 0%      5.00 ± 0%  -37.50%  (p=0.000 n=10+10)